Add some more codec mappings and rename for more clarity
Video codec id 2 is not actually real H263 but Sorenson H263, aka Flash
Video version 1. Rename to SORENSON_H263.

Video codec id 8 and 9 are not specified in the standard, but are used
by files out in the wild on supported by FFMPEG and GStreamer. They're
(real) H263 and MPEG-4 Part 2.

Audio codec id 0 is not big endian PCM, but it is defined as native
endianness of the system that created the stream. Which is of course a
rather useless specification. Rename it from PCM_BE to PCM_NE.